Move authapp and unauthapp into app containers
