+++ app/src/components/AppContainers/UnauthenticatedApp.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../Header/UnauthHeader";
 import Footer from "../Footer/Footer";
-import AuthContainer from "./AuthContainer";
+import AuthContainer from "../Authentication/AuthContainer";
 
 const UnauthenticatedApp = () => {
   return (
